Factor out json handling
